requirements.txt
msal
requests
python-dotenv
langchain
docling
langchain-community
langchain_docling
langchain-openai
langchain-milvus==0.1.10
pymilvus
flask
flask-cors
firecrawl-py
langchain-core
gunicorn